unittest: fix all unsigned vs. signed comparisons
gcc 5.3.1 breaks compilation due to all these unsigned vs.
signed comparisons.  On previous gcc versions these should
be non-fatal warnings

TEST="compile with --enable-gtest and gcc 5.3.1"

+++ common/nalreader_unittest.cpp
@@ -77,7 +77,7 @@ NALREADER_TEST(ReadAsWhole) {
     NalReader reader(&g_data[0], g_data.size(), true);
 
     EXPECT_TRUE(reader.read(nal, size));
-    EXPECT_EQ(size, g_data.size());
+    EXPECT_EQ((size_t)size, g_data.size());
 
     for (int32_t i(0); i < size; ++i)
         EXPECT_EQ(nal[i], g_data[i]);
